defect(apps): searching for SAM apps does not work
* it should return superAdminApp
* it should NOT return userApp NOR adminApp

This commit fixes it.

Closes [RUNTIME-331](https://bonitasoft.atlassian.net/browse/RUNTIME-331)

+++ bonita-integration-tests/bonita-test-utils/src/main/java/org/bonitasoft/engine/test/APITestUtil.java
@@ -1351,7 +1351,9 @@ public class APITestUtil extends PlatformTestUtil {
         final SearchResult<Application> applications = getApplicationAPI()
                 .searchApplications(new SearchOptionsBuilder(0, 100).done());
         for (Application application : applications.getResult()) {
-            getApplicationAPI().deleteApplication(application.getId());
+            if (application.isEditable()) {
+                getApplicationAPI().deleteApplication(application.getId());
+            }
         }
     }
